reorganize files and split components
