+++ packages/extension/src/ui/features/actions/transaction/ApproveTransactionScreen/ConfirmScreen.tsx
@@ -4,12 +4,12 @@ import { FC, ReactNode, useState } from "react"
 import Measure from "react-measure"
 import { useNavigate } from "react-router-dom"
 
-import { formatTruncatedAddress } from "../../services/addresses"
-import { Account } from "../accounts/Account"
+import { formatTruncatedAddress } from "../../../../services/addresses"
+import { Account } from "../../../accounts/Account"
 import {
   getAccountName,
   useAccountMetadata,
-} from "../accounts/accountMetadata.state"
+} from "../../../accounts/accountMetadata.state"
 
 export interface ConfirmPageProps {
   onSubmit?: (e: React.FormEvent<HTMLFormElement>) => void
