+++ packages/extension/src/ui/features/actions/transaction/ApproveTransactionScreen/DappHeader/TransactionIcon/DeclareTransactionIcon.tsx
@@ -1,0 +1,11 @@
+import { DocumentIcon } from "@argent/ui/dist/components/icons"
+
+import { IconWrapper } from "./IconWrapper"
+
+export const DeclareContractIcon = () => {
+  return (
+    <IconWrapper position="relative" borderRadius="90">
+      <DocumentIcon />
+    </IconWrapper>
+  )
+}
