+++ packages/extension/src/ui/features/actions/transaction/ApproveTransactionScreen/VerifiedDappBanner/VerifiedDappModal.tsx
